backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Nicolas Kasevich <n.kasevich@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.0"}
alembic = "^1.11.3"
psycopg2-binary = "^2.9.7"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}



[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
coverage = "^7.3.0"
mock = "^5.1.0"
pytest-pretty = "^1.2.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.coverage.run]
omit = [
    ".venv/*",
    "alembic/*",
]